Moved root namespace to directory.build.props file Renamed Inventory to Inventory.Api Synced namespaces
