+++ src/Inventory.Api/AutoGen/swagger.json
