+++ src/Inventory.Api/Properties/launchSettings.json
