+++ src/Inventory.Api/appsettings.Development.json
