+++ src/Inventory.Api/appsettings.json
