+++ src/Inventory.Data.Abstractions/Inventory.Data.Abstractions.csproj
@@ -1,11 +1,5 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <AssemblyName>TheVideoGameStore.Inventory.Data.Abstractions</AssemblyName>
-    <RootNamespace>TheVideoGameStore.Inventory.Data.Abstractions</RootNamespace>
-  </PropertyGroup>
-
   <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.0" />
   </ItemGroup>
