+++ src/Inventory.Data.Models/Inventory.Data.Models.csproj
@@ -1,9 +1,3 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <AssemblyName>TheVideoGameStore.Inventory.Data.Models</AssemblyName>
-    <RootNamespace>TheVideoGameStore.Inventory.Data.Models</RootNamespace>
-  </PropertyGroup>
-
 </Project>
