+++ src/Inventory.Models/Inventory.Models.csproj
@@ -1,9 +1,3 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <AssemblyName>TheVideoGameStore.Inventory.Models</AssemblyName>
-    <RootNamespace>TheVideoGameStore.Inventory.Models</RootNamespace>
-  </PropertyGroup>
-
 </Project>
