+++ src/Inventory.Repositories.Tests/Inventory.Repositories.Tests.csproj
@@ -1,12 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-
     <IsPackable>false</IsPackable>
-
-    <RootNamespace>TheVideoGameStore.$(MSBuildProjectName.Replace(" ", "_"))</RootNamespace>
 
     <PackageId>TheVideoGameStore.$(AssemblyName)</PackageId>
   </PropertyGroup>
