+++ src/directory.build.props
@@ -1,0 +1,7 @@
+<Project>
+  <PropertyGroup>
+    <AssemblyName>TheVideoGameStore.$(MSBuildProjectName)</AssemblyName>
+    <RootNamespace>TheVideoGameStore.$(MSBuildProjectName.Replace(" ", "_"))</RootNamespace>
+    <TargetFramework>net6.0</TargetFramework>
+  </PropertyGroup>
+</Project>
